feat: land current attendance delivery snapshot
Sync the working tree covering report queries, paper overtime OCR,
punch exemption, Deli replay, leave-account OA recalc, identity
cutover, and related OpenSpec/docs. Include the performance
optimization task list as the baseline for the next pass.

+++ backend/src/main/java/com/szsemicon/hr/reporting/infrastructure/persistence/AttendanceReportProjectionWriteRows.java
@@ -108,7 +108,7 @@ final class AttendanceReportProjectionWriteRows {
             long absenceMinutes,
             long actualWorkMinutes,
             int scheduledAttendanceDays,
-            int actualAttendanceDays,
+            double actualAttendanceDays,
             long lateMinutes,
             long penalizedLateMinutes,
             long earlyDepartureMinutes,
